Header UI Refactoring and Organization
-Fix the header position and add shadow and border to clearly separate it visually.
-Align the logo, “Create” button, child selector, and notification bell to the right.
-Add a dropdown to the “Create” button to consolidate links for various record creation.
-Standardize the child selector UI using a form + button, and improve display when no child is selected.
-Adjust styles around the notification bell and improve badge display.
-Clean up CSS and unify margins with the main content and sidebar.
-Modularize JS, separating header and notification toggle functionalities.
