+++ app/assets/stylesheets/application.css
@@ -79,6 +79,14 @@ select {
 }
 
 /* ==============================
+   mainコンテンツ領域
+   ============================== */
+.main-content {
+  margin-top: 60px; /* ヘッダー分の余白を確保 */
+  padding: 20px;    /* 中身と枠の余白（任意） */
+}
+
+/* ==============================
    フィルターバー（一覧ページ上部の操作用UI）
    ============================== */
 .filter-bar {
@@ -95,30 +103,65 @@ select {
    ヘッダー共通
    ============================== */
 .site-header {
+  position: fixed;
+  top: 0;
+  left: 0;
+  width: 100%;
   background-color: #f8f8f8;
   padding: 10px 20px;
-  width: 100%;
-  overflow-x: hidden; /* 横スクロール防止 */
+  overflow: visible;
+  z-index: 1100;
+  border-bottom: 1px solid #ddd; /* ← 境界線を追加 */
+  box-shadow: 0 1px 2px rgba(0,0,0,0.05); /* ← 薄い影で立体感 */
 }
 
-.header-nav {
+.header-container {
   display: flex;
-  flex-wrap: wrap; /* 必要に応じて折り返す */
-  gap: 8px;
-  max-width: 100%;
+  justify-content: space-between;
+  align-items: center;
 }
 
-.nav-btn {
-  display: inline-block;
-  padding: 6px 12px;
-  background-color: #ffffff;
-  border: 1px solid #ccc;
-  border-radius: 4px;
+.header-logo .logo-text {
+  font-size: 20px;
+  font-weight: bold;
   text-decoration: none;
   color: #333;
+}
+
+.header-actions {
+  display: flex;
+  align-items: center;
+  gap: 20px;
+}
+
+.child-status {
   font-size: 14px;
-  text-align: center;
-  box-sizing: border-box;
+  color: #333;
+}
+
+.notification-link {
+  position: relative;
+  color: #333;
+}
+
+/* 通知ベルアイコン */
+.notification-link i.fa-bell {
+  font-size: 24px;       /* ベルの大きさを調整 */
+  line-height: 1;        /* バッジとの縦位置調整用 */
+  vertical-align: middle;
+}
+
+/* 通知カウントバッジ */
+.notification-count {
+  position: absolute;
+  top: -6px;
+  right: -10px;          /* ベルが大きくなった分少し右に */
+  background: red;
+  color: white;
+  font-size: 14px;       /* バッジも少し大きく */
+  font-weight: bold;
+  padding: 3px 7px;      /* バッジ内余白を広げる */
+  border-radius: 50%;
 }
 
 /* ==============================
@@ -127,6 +170,85 @@ select {
 #user-menu {
   width: auto;
   max-width: 95%;
+}
+
+/* ==============================
+   アカウントメニュー共通スタイル
+   ============================== */
+#account-menu a,
+#account-menu form,
+#account-menu button {
+  display: block;
+  width: 100%;
+  padding: 8px 12px;      /* 内側余白 */
+  margin-bottom: 4px;      /* アイテム間隔 */
+  text-align: left;
+  background-color: white;
+  border: none;
+  border-radius: 0;
+  color: #333;
+  text-decoration: none;
+  cursor: pointer;
+}
+
+#account-menu a:hover,
+#account-menu form:hover,
+#account-menu button:hover {
+  background-color: #f0f0f0; /* ホバーで少し色を変えるだけ */
+}
+
+/* ==============================
+   作成ボタン・ドロップダウン共通
+   ============================== */
+#new-record-button {
+  padding: 6px 12px;
+  font-size: 16px;
+  background-color: #1d4ed8;  /* ブルー */
+  color: #fff;
+  border: none;
+  border-radius: 6px;
+  cursor: pointer;
+  transition: background-color 0.2s;
+}
+
+#new-record-button:hover {
+  background-color: #2563eb;
+}
+
+/* ドロップダウンメニュー */
+#record-icons {
+  position: absolute;
+  top: 100%;
+  right: 0;
+  background-color: #fff;
+  border: 1px solid #e5e7eb; /* 薄いグレーで境界 */
+  border-radius: 6px;
+  box-shadow: 0 4px 8px rgba(0,0,0,0.1);
+  display: flex;
+  flex-direction: column;
+  gap: 4px;
+  padding: 8px;
+  min-width: 120px;
+  z-index: 1200;
+}
+
+/* ドロップダウン内リンク */
+#record-icons a {
+  display: block;
+  padding: 6px 10px;
+  color: #1f2937;  /* ダークグレー */
+  text-decoration: none;
+  border-radius: 4px;
+  transition: background-color 0.2s;
+}
+
+#record-icons a:hover {
+  background-color: #f3f4f6; /* 薄いグレー */
+}
+
+/* ドロップダウン非表示状態 */
+.hidden {
+  display: none !important;
 }
 
 /* ==============================
