+++ app/assets/stylesheets/sidebar.css
@@ -2,15 +2,17 @@
    サイドバー専用スタイル
    ============================== */
 
+
 /* デフォルト非表示で左に隠す */
 .sidebar {
   position: fixed;
-  top: 0;
-  left: -260px; /* 隠しておく */
+  top: 60px; /* ヘッダーの高さ分ずらす */
+  left: -260px;
   width: 240px;
-  height: 100%;
+  height: calc(100% - 60px);
   background-color: #f9f9f9;
-  border-right: 1px solid #ddd;
+  border-right: 1px solid #ddd; /* ← ヘッダーと同じ線を追加 */
+  box-shadow: none; /* ← 影は消す（線に統一） */
   padding: 16px;
   font-size: 14px;
   overflow-y: auto;
@@ -156,5 +158,24 @@
 
   .notification-setting {
     padding: 8px;
+  }
+}
+
+/* ==============================
+   PC常時表示
+   ============================== */
+@media screen and (min-width: 1025px) {
+  .sidebar {
+    left: 0;          /* 常時表示 */
+    transition: none; /* アニメーション不要 */
+  }
+
+  .sidebar-overlay {
+    display: none !important;
+  }
+
+  .main-content {
+    margin-left: 240px; /* サイドバー分ずらす */
+    margin-top: 60px;   /* ヘッダー分の余白を確保 */
   }
 }
